Memory card code adjustments
Changed fileInfo to be a local variable on the stack instead of allocating memory.

Changed workArea to use MemCardWorkArea0, since that's what the vanilla code uses for the work area variable.

CARDProbeEx can return CARD_RESULT_BUSY, so made it look until it doesn't return that value. To prevent an infinite loop, it will run a maximum of 1,000,000 times.

Fixed some formatting issues.