Adjust callRelProlog and remove callRelPrologMounted
callRelProlog now properly handles mounting/unmounting at the proper times.

Instead of allocating a temporary buffer to load the REL file from the memory card, memmove is now used to copy the REL file to the start of the buffer. This helps to avoid a lot of unnecessary memory being used at once if loading a very large REL file.